requirements.txt
python-hcl2==4.3.0
PyYAML==6.0.1
pytest==8.0.0
